Move everything into Explorer.js following review
