--- src/plugins/graphql-scalar-inputs/bundled/DateInput.js
@@ -1,24 +1,0 @@
-import * as React from 'react';
-
-function canProcess(arg) {
-  return arg && arg.type && arg.type.name ===  'Date';
-}
-
-function render(props) {
-  return (
-    <input
-      type="date"
-      value={props.arg.defaultValue}
-      onChange={props.setArgValue}
-    />
-  );
-}
-
-
-const DatePlugin = {
-  canProcess,
-  render,
-  name: 'DateInput'
-};
-
-export default DatePlugin;
